Updating auth

+++ backend/hus_bakery_app/routers/admin/coupon_management.py
@@ -1,4 +1,6 @@
+import json
 from flask import Blueprint, request, jsonify
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from hus_bakery_app.services.admin.coupon_management_services import (
     get_all_coupons_service, add_coupon_service,
     edit_coupon_service, delete_coupon_service
@@ -7,7 +9,12 @@ from hus_bakery_app.services.admin.coupon_management_services import (
 coupon_admin_bp = Blueprint('coupon_admin_bp', __name__)
 
 @coupon_admin_bp.route('/coupon', methods=['GET'])
+@jwt_required()
 def get_coupons():
+    identity = json.loads(get_jwt_identity())
+    if identity.get("role") != 'employee':
+        return jsonify({"error": "Bạn không có quyền truy cập dữ liệu này"}), 403
+
     status_filter = request.args.get('status')
     raw_coupons = get_all_coupons_service()
 
@@ -35,7 +42,12 @@ def get_coupons():
 
 
 @coupon_admin_bp.route('/add_coupon', methods=['POST'])
+@jwt_required()
 def add_coupon():
+    identity = json.loads(get_jwt_identity())
+    if identity.get("role") != 'employee':
+        return jsonify({"error": "Bạn không có quyền thực hiện thao tác này"}), 403
+
     data = request.json
     try:
         new_coupon = add_coupon_service(data)
@@ -45,7 +57,12 @@ def add_coupon():
 
 
 @coupon_admin_bp.route('/update_coupon/<int:coupon_id>', methods=['PUT'])
+@jwt_required()
 def update_coupon(coupon_id):
+    identity = json.loads(get_jwt_identity())
+    if identity.get("role") != 'employee':
+        return jsonify({"error": "Bạn không có quyền chỉnh sửa dữ liệu"}), 403
+
     data = request.json
     if edit_coupon_service(coupon_id, data):
         return jsonify({"message": "Cập nhật thành công"}), 200
@@ -53,7 +70,12 @@ def update_coupon(coupon_id):
 
 
 @coupon_admin_bp.route('/delete_coupon/<int:coupon_id>', methods=['DELETE'])
+@jwt_required()
 def delete_coupon(coupon_id):
+    identity = json.loads(get_jwt_identity())
+    if identity.get("role") != 'employee':
+        return jsonify({"error": "Bạn không có quyền xóa dữ liệu"}), 403
+
     if delete_coupon_service(coupon_id):
         return jsonify({"message": "Xóa mã giảm giá thành công"}), 200
     return jsonify({"error": "Không tìm thấy mã giảm giá"}), 404
